Implement base class to be able to hide complexity
